create benchmarks project to enable dedicated running on benchmarks
