+++ harmony.sln
@@ -21,6 +21,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		src\.editorconfig = src\.editorconfig
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SIL.Harmony.Benchmarks", "src\SIL.Harmony.Benchmarks\SIL.Harmony.Benchmarks.csproj", "{96FABEBB-5A18-46D5-8530-84B9FFF5442C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -54,5 +56,9 @@ Global
 		{8EB807F6-C548-4016-856E-2ACCA5603036}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8EB807F6-C548-4016-856E-2ACCA5603036}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8EB807F6-C548-4016-856E-2ACCA5603036}.Release|Any CPU.Build.0 = Release|Any CPU
+		{96FABEBB-5A18-46D5-8530-84B9FFF5442C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{96FABEBB-5A18-46D5-8530-84B9FFF5442C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{96FABEBB-5A18-46D5-8530-84B9FFF5442C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{96FABEBB-5A18-46D5-8530-84B9FFF5442C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
